chore: Port to 1.21, using NeoForge EventBus (#2597)
* port: Change version to 1.21

* port: Use Java 21 in Github CI

* port: partialTicks -> deltaTracker changes

* port: Misc mixin fixes

* port: ResourceLocation ctor -> factory methods

* port: ClientCommandManager remove deprecated error

* port: BufferBuilder -> ByteBufferBuilder

* port: Port RenderUtils and BufferedRenderUtils (VertexBuilder changes)

* port: Port FontRenderer, MapRenderer and LootrunRenderer (VertexBuilder changes)

* port: ItemStack#getTooltipLines related changes

* port: Port BeaconModel (Entity -> LivingEntity)

* port: Port EntityDataAccessor users (getId -> id)

* port: Port ServerboundUseItemPacket usages

* port: Port WynncraftButtonFeature (fix ConnectScreen#startConnecting arguments, pingServer arguments)

* port: Port BeaconBeamFeature (use Minecraft's BeaconRenderer, remove custom renderer)

* port: Port to using DataComponents

* port: Port AccessWidener changes

* port: Port ClientCommandManager (CommandBuildContext changes)

* port: Port ScoreboardHandler (PlayerScores/ScoreHolder changes)

* port: Port ResourcePackService

* port: Port to 1.21 NeoForge (buildscript, mixins)

This commit ports to 1.21 NeoForge. Notable changes are the fixed NeoForge/Fabric only mixins (I've separated old, require = 0 mixins to Fabric, so it's easier to see what works where). I've also added improvements to our buildscript for both Fabric/NeoForge, to aid version changes.

This commit does not compile, as NeoForge's new EventBus needs some changes to the code base.

* port: Port to NeoForge EventBus

This PR is mostly just import changes. A few rules were added to NeoForge's event bus, to make it easier and less error-prone to use.
Affected parts:
- @Cancelable annotation is now an interface.
- With the annotations removed, the event bus does not need pre-processing, so we no longer depend on EventBusTransformer.
- Generic events were removed. Only PackedEvent used it, it's a simple port.
- Abstract classes cannot register events with @SubscribeEvent. For overlays, I've introduced manager level ticking. For special overlays, I've added WynntilsMod#registerListener to specially register these events.
- Abstract events cannot be listened to. This resulted in a small amount of code duplication. Not enough to consider doing any refactors.

* port: Port CI, README.md to NeoForge

* port: Port RenderUtils#renderCustomNametag (yes, really.)

Between 1.20.2 and 1.21, they removed the - sign from the x coordinate, and it works. Don't ask me.

* fix: Patch modmenu compatibility

* chore: Bump version to 2.0.0-SNAPSHOT

* fix: Fix LootrunRenderer rendering (LevelRenderer PoseStack changes)

* fix: Fix failing tests due to Minecraft's Style class changes

+++ common/src/main/java/com/wynntils/mc/event/MobEffectEvent.java
@@ -1,12 +1,12 @@
 /*
- * Copyright © Wynntils 2022-2023.
+ * Copyright © Wynntils 2022-2024.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.mc.event;
 
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.entity.Entity;
-import net.minecraftforge.eventbus.api.Event;
+import net.neoforged.bus.api.Event;
 
 public abstract class MobEffectEvent extends Event {
     private final Entity entity;
@@ -26,16 +26,16 @@ public abstract class MobEffectEvent extends Event {
     }
 
     public static final class Update extends MobEffectEvent {
-        private final byte effectAmplifier;
+        private final int effectAmplifier;
         private final int effectDurationTicks;
 
-        public Update(Entity entity, MobEffect effect, byte effectAmplifier, int effectDurationTicks) {
+        public Update(Entity entity, MobEffect effect, int effectAmplifier, int effectDurationTicks) {
             super(entity, effect);
             this.effectAmplifier = effectAmplifier;
             this.effectDurationTicks = effectDurationTicks;
         }
 
-        public byte getEffectAmplifier() {
+        public int getEffectAmplifier() {
             return effectAmplifier;
         }
